// bikematrix.frontend/generate/models/BikeDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BikeMatrix API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { BrandDTO } from './BrandDTO';
import {
    BrandDTOFromJSON,
    BrandDTOFromJSONTyped,
    BrandDTOToJSON,
} from './BrandDTO';
import type { OwnerDTO } from './OwnerDTO';
import {
    OwnerDTOFromJSON,
    OwnerDTOFromJSONTyped,
    OwnerDTOToJSON,
} from './OwnerDTO';

/**
 * 
 * @export
 * @interface BikeDTO
 */
export interface BikeDTO {
    /**
     * 
     * @type {number}
     * @memberof BikeDTO
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof BikeDTO
     */
    model?: string | null;
    /**
     * 
     * @type {number}
     * @memberof BikeDTO
     */
    year?: number;
    /**
     * 
     * @type {OwnerDTO}
     * @memberof BikeDTO
     */
    owner?: OwnerDTO;
    /**
     * 
     * @type {BrandDTO}
     * @memberof BikeDTO
     */
    brand?: BrandDTO;
}

/**
 * Check if a given object implements the BikeDTO interface.
 */
export function instanceOfBikeDTO(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function BikeDTOFromJSON(json: any): BikeDTO {
    return BikeDTOFromJSONTyped(json, false);
}

export function BikeDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): BikeDTO {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': !exists(json, 'id') ? undefined : json['id'],
        'model': !exists(json, 'model') ? undefined : json['model'],
        'year': !exists(json, 'year') ? undefined : json['year'],
        'owner': !exists(json, 'owner') ? undefined : OwnerDTOFromJSON(json['owner']),
        'brand': !exists(json, 'brand') ? undefined : BrandDTOFromJSON(json['brand']),
    };
}

export function BikeDTOToJSON(value?: BikeDTO | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'model': value.model,
        'year': value.year,
        'owner': OwnerDTOToJSON(value.owner),
        'brand': BrandDTOToJSON(value.brand),
    };
}